+++ tools/hotpatch-release.py
@@ -23,17 +23,27 @@
 # ---------------------------------------------------------------------------------
 
 from urllib.request import Request, urlopen
+from pathlib import Path
+from io import StringIO
 
+import hashlib
+import base64
 import json
 import os
 
-token = os.environ["GITHUB_TOKEN"] or ""
-tag = os.environ["GITHUB_REF_NAME"] or ""
+REPOSITORY = "Noelware/violet"
 
-def _github_request(url: str, method: str = "GET", data = None):
+token = os.environ.get("GITHUB_TOKEN")
+tag = os.environ.get("GITHUB_REF_NAME")
+wd = os.environ.get("GITHUB_WORKSPACE")
+
+if not wd:
+    wd = str(Path.cwd())
+
+def GitHubRequest(url: str, method: str = "GET", data = None):
     headers = {
         "Accept": "application/vnd.github+json",
-        "Authorization": f"token {token}"
+        "Authorization": "token %s" % token
     }
 
     if data is not None:
@@ -44,48 +54,79 @@ def _github_request(url: str, method: str = "GET", data = None):
     with urlopen(req) as resp:
         return json.load(resp)
 
-def main():
-    if len(token) == 0 or len(tag) == 0:
+def MesonHotpatch(f: StringIO, tag: str):
+    mapping = {
+        "violet": "violet_dep",
+        "violet_experimental": "violet_experimental_dep",
+        "violet_io": "violet_io_dep",
+        "violet_filesystem": "violet_filesystem_dep",
+        "violet_experimental_threading": "violet_experimental_threading_dep",
+        "violet_io_experimental": "violet_io_experimental_dep",
+        "violet_terminal": "violet_terminal_dep",
+        "violet_anyhow": "violet_anyhow_dep"
+    }
+
+    with open("%s/mesondist.tgz" % wd, "rb") as fd:
+        source_hash = hashlib.sha256(fd.read()).hexdigest()
+
+    f.write('## Meson\n')
+    f.write('> **subprojects/violet.wrap**:\n')
+    f.write('```ini\n')
+    f.write('[wrap-file]\n')
+    f.write('source_url  = https://github.com/%s/releases/download/%s/mesondist.tgz\n' % (REPOSITORY, tag))
+    f.write('source_hash = %s\n\n' % source_hash)
+    f.write('[provide]\n')
+
+    for name, dep in mapping.items():
+        f.write('%s = %s\n' % (name, dep))
+
+    f.write('```\n\n')
+
+def BazelHotpatch(f: StringIO, tag: str):
+    with open("%s/bazeldist.tgz" % wd, "rb") as fd:
+        digest = hashlib.sha256(fd.read()).digest()
+
+    integrity = "sha256-%s" % base64.b64encode(digest).decode()
+    f.write('## Bazel\n')
+    f.write('> As of **21/01/26**, we plans on uploading all **Noelware.Violet** frameworks onto the BCR and our own registry (`https://bzl.noelware.cloud`). If either are available, prefer using `--registry=https://bzl.noelware.cloud`.\n')
+    f.write('\n```python\n')
+    f.write('bazel_dep(name = "violet", version = "%s")\n' % tag)
+    f.write('archive_override(\n')
+    f.write('    module_name = "violet",\n')
+    f.write('    integrity = "%s",\n' % integrity)
+    f.write('    urls = [\n')
+    f.write('          # "https://artifacts.noelware.org/bazel-registry/violet/%s/bazeldist.tgz",\n' % tag)
+    f.write('          "https://github.com/%s/releases/download/%s/bazeldist.tgz",\n' % (REPOSITORY, tag))
+    f.write('    ]\n')
+    f.write(')\n\n')
+
+def Main():
+    if token is None or tag is None == 0:
         raise RuntimeError("$GITHUB_TOKEN or $GITHUB_REF_NAME not defined in environment")
 
-    release = _github_request(f"https://api.github.com/repos/Noelware/violet/releases/tags/{tag}")
+    release = GitHubRequest("https://api.github.com/repos/%s/releases/tags/%s" % (REPOSITORY, tag))
     body = release["body"]
 
-    print("==== current release notes ====")
+    print("=== current release notes ====")
     print(body)
-    print("====                       ====")
+    print("===                       ====")
 
-    if "BAZELDIST_INTEGRITY" in os.environ:
-        integrity = os.environ.get("BAZELDIST_INTEGRITY")
+    f = StringIO()
+    f.write(body)
 
-        body += "\n\n## Bazel\n"
-        body += "> As of **21/01/26**, we have plans on uploading all **Noelware.Violet** frameworks onto the BCR and our own (`https://bzl.noelware.cloud`). It could exist right now, if you are, you can use `--registry=https://bzl.noelware.cloud` and all frameworks will be available."
-        body += "\n"
-        body += "\n```python\n"
-        body += "bazel_dep(name = \"violet\", version = \"%s\")\n" % tag
-        body += "archive_override(\n"
-        body += "    module_name = \"violet\",\n"
-        body += "    integrity = \"%s\",\n" % integrity
-        body += "    urls = [\n"
-        body += "        # \"https://artifacts.noelware.org/bazel-registry/violet/%s/bazeldist.tgz\",\n" % tag
-        body += "        \"https://github.com/Noelware/violet/releases/download/%s/bazeldist.tgz\",\n" % tag
-        body += "    ]\n"
-        body += ")\n"
-        body += "```\n\n"
+    BazelHotpatch(f, tag)
+    MesonHotpatch(f, tag)
 
-    # TODO: Meson, CMake hotpatch
-    body += "## CMake\n"
-    body += "Right now, it is not the best to use the frameworks in a CMake environment! *ehehe...~*\n"
-    body += "\n## Meson \n"
-    body += "Using the frameworks in a Meson environment is not ready yet!\n"
-    body += "\n## GN (\"Generate Ninja\")\n"
-    body += "Using the frameworks in a GN environment is not ready yet!\n"
+    f.write('## CMake\n')
+    f.write('As of this release, CMake distribution is not available~\n\n')
+    f.write('## GN\n')
+    f.write('As of this release, GN distribution is not available~\n\n')
 
-    _github_request(
-        f"https://api.github.com/repos/Noelware/violet/releases/{release["id"]}",
+    GitHubRequest(
+        "https://api.github.com/repos/%s/releases/%s" % (REPOSITORY, release["id"]),
         method="PATCH",
-        data={"body": body}
+        data={"body": f.getvalue()}
     )
 
 if __name__ == '__main__':
-    main()
+    Main()